Fix exception when finding relationship to home person
Proposed bug fix to catch the reported exception by testing for None.
However, it's not a guaranteed fix because no data was provided to
test the bug.

Question remains what is the family tree that leads to that situation?
Is it a valid case? Should it be handled in a certain way that isn't
being done now? If it isn't valid, is it being captured by check and
repair database?

Fixes #13495