New indexes for big databases in the narrative web report
When we have large databases, the current indexes are too big.
We now have a single page for each letter.

We can limit the size of the page in rows per page in the html tab of
the report. The values are between 10 and 2000. The default is 500.
If we have more than this limit, we create a secondary index for the
associated letter.