Re-submit after refactoring
Hi Nick,
I refactored the creation and invocation of styles by gender and family boxes to the utils class.  I am still using style sheets to control the box color.  Is it possible to consider this as a stop-gap solution, for the time being?

I am also looking at using graph reports such as family-line and hour-glass.  It would be a major rewrite, in my estimation.  I like their extra capability, and I am thinking of creating graph-based reports for ancestors and descendants that allow "scale to fit page width only", which is important for me, considering the huge number of nodes for my family.  But this is going to take a lot of research and coding.  I am not confident that I can do it on my own.