Extending the same feature to the Ancestor Report
Even though this is a different report, it is also dependent on the same report/utils class, as the result of refactoring the function to handle the color of the boxes.  The code changes are similar between the 2 reports.  I am requesting that this feature be extended to this report as well.  It will make it easier for the reviewing and approval process of the PR easier than having to deal with two.