Add the Calendar to Person, Family and Event
For the active person, if you use the "Person Calendar" gramplet,
the birth date will be shown on the calendar.

For the active family, if you use the "Family Calendar" gramplet,
the marriage date will be shown on the calendar.

For the active event, if you use the "Event Calendar" gramplet,
the event date will be shown on the calendar.

You can always us the calendar in the dashboard.
The double click will do the same things as before.