Dev version correctly reports the SHA of the local HEAD commit
Local builds report the Gramps version in the form of Major.Minor.Patch-SHA. However, the commit SHA reported did not correspond to the repo root, but that of the gramps/gramps folder which was unexpected. This change fixes that by removing the path argument to get_git_revision so the git log command does not limit commits, and the dev version is reported as expected.

A brief history: the path argument was a carry-over from SVN, and wasn't initially used (see commit 57daf4f), but a subsequent change attempted to use it (commit 6836dbd), and finally the path argument was appended to the git log command in commit f92ee32.

To test this change, run `python Gramps.py -v` from the root of the source directory, or other subfolders (adjusting the path to Gramps.py as required) and verify that the version reported matches the HEAD commit of your source.

Tested in an MSYS environment on Windows, and Ubuntu (in a WSL environment). Please verify on Apple MacOS, and a Linux system.