Suppress file not found error if not available at end of gpkg import
It's not clear why data.gramps file isn't available in certain situations as
reported by users. In examining the code however, there were no errors opening
or importing the data, and exception is thrown after the data has already been
imported. Given that we have the data it is inconsequential that the file isn't
found when removing it, so the exception can be suppressed. If there is a
different exception that occurs, that will still be shown to the user which we
can investigate if/when reported.

Fixes #12311.