Add tearDownClass method to cleaup after unit test
test_editreference created a database in /tmp but did not clean up
after the test was done. This causes a test failure when the database
schema changes and a newer version of Gramps runs the tests. The
test will fail with an upgrade required message.

Now the test closes the database after it is done, and a tearDownClass
method is added which removes the database folder.