Select next item in list instead of most recently accessed
After deleting a database, the list gets rebuilt and the selected items defaults to the most recently accessed list item. This usually results in the selection jumping around to a completely different place in the list which is jarring.

This change will remember the item that was next in the selection (previous if there is no next), and select that after the current items is deleted.

Fixes #12187